Fix parsing of lowercase S/W directions
We were matching the direction case-insensitively in the regex, but then
comparing it case-sensitively afterwards, so that lowercase S/W
coordinates would not be parsed as S/W and would instead eventually end
up being interpreted as unsigned (i. e. N/E). To fix this, it’s
sufficient to check if the regex matched that part of the pattern at
all or fell back to the empty alternative branch.

Credit for investigating the bug goes to Wikidata user Nikki [1] on
Phabricator [2].

This is a backport for the 3.0.x line, cherry-picked from 54437223f4.

[1]: https://www.wikidata.org/wiki/User:Nikki
[2]: https://phabricator.wikimedia.org/T173026#4465312

Bug: T173026