Upgrade mediawiki-codesniffer to latest version (45.0.0)
This repository is currently using version 34 of the Mediawiki coding
style phpcs rules. If the code aims to comply with Mediawiki style
guidelines, it makes sense that it continues to comply at current
versions of the rules.

Besides being simply different, the newer versions of the rules
introduce, for example,
MediaWiki.Usage.NullableType.ExplicitNullableTypes, which enforces
compliance with coming deprecation rules in PHP 8.4 concerning nullable
types. If the code does not remove the soon-to-be-deprecated form of
nullable type declarations, the library will no longer be usable in
Mediawiki for PHP 8.4 deployments.

Note that mediawiki-codesniffer 45.0.0 requires PHP version 7.4 or
later - PHP 7.3 support will be dropped with this change. PHP 7.3
has been out of support since 6. Dec 2021.

Resolves #222

Bug: T379481